Moved QTI version classes to data\storage\xml namespace.
