+++ src/qtism/data/storage/xml/versions/CompactVersion22.php
@@ -19,17 +19,16 @@
  * @author Julien Sébire <julien@taotesting.com>
  * @license GPLv2
  */
-namespace qtism\common\utils\versions;
+namespace qtism\data\storage\xml\versions;
 
 /**
- * QTI version 2.2.2
+ * QTI Compact version 2.2
  */
-class QtiVersion222 extends QtiVersion220
+class CompactVersion22 extends CompactVersion
 {
-    const XSD = 'http://www.imsglobal.org/xsd/qti/qtiv2p2/imsqti_v2p2p2.xsd';
+    const XMLNS = 'http://www.imsglobal.org/xsd/imsqti_v2p2';
 
-    public function getSchemaLocation()
-    {
-        return 'qtiv2p2p2/imsqti_v2p2p2.xsd';
-    }
+    const XSD = 'http://www.taotesting.com/xsd/qticompact_v2p2.xsd';
+
+    const LOCAL_XSD = 'qticompact_v2p2.xsd';
 }
